refactor: move lib/gorm -> infrastructure/gorm
