+++ src/infrastructure/gorm.go
@@ -1,4 +1,4 @@
-package lib
+package infrastructure
 
 import (
 	"fmt"
